feat: implement record methods
Being able to create stubs manually is great, but when you work with the
SDK of a SAAS, you are not supposed to know what goes on the wire. Being
able to record that, put it under version control and be done with it
allows to remove that information from tests and focus on functional
assertions.