Fix new methods get/find/delete requests
Multiple fixes are included in this commit.

First, the FindRequestsByCriteria method now actually calls the /find
endpoint, not the /remove one.

Second, the new structs to deserialize the responses from the journal
requests were reviewed to better match what is actually returned by
wiremock. Some fields might be missing, but the ones that are present
should be correct. Any missing field can be added in the future without
breaking the api.

Lastly, the methods have been renamed to better match the api
documentation, and to enforce golang conventions.