Use `strings` instead of `[]byte` for Body fields
The body fields of Request and Response may contain binary data, but it
is not base64 encoded, so json.Unmarshal can't process it to `[]byte`.
We therefore have no choice than to use a `string` for that. It is up to
the user to decide if they want to use that value directly or to
deserialize the base64 encoded BodyAsBase64.